Model performance plot

This function generates plots to visualize model performance

Usage

performance_plot(
  model_list,
  type = "box",
  text_size = 10,
  palette = "viridis",
  save = FALSE,
  file_name = "Performance_plot",
  file_type = "pdf",
  plot_width = 7,
  plot_height = 7,
  dpi = 80
)

Arguments

model_list

A model_list object from performing train_models.

type

Type of plot to generate. Choices are "box" or "dot." Default is "box." for boxplots.

text_size

Text size for plot labels, axis labels etc. Default is 10.

palette

Viridis color palette option for plots. Default is "viridis". See viridis for available options.

save

Logical. If TRUE saves a copy of the plot in the working directory.

file_name

file_name File name to save the plot. Default is "Performance_plot."

file_type

File type to save the plot. Default is "pdf".

plot_width

Width of the plot. Default is 7.

plot_height

Height of the plot. Default is 7.

dpi

Plot resolution. Default is 80.

Value

A ggplot2 object.

Details

The default metrics used for classification based models are "Accuracy" and "Kappa."

These metric types can be changed by providing additional arguments to the train_models function. See train and trainControl for more information.

See also

Author

Chathurani Ranathunge

Examples

if (FALSE) {

## Create a model_df object
covid_model_df <- pre_process(covid_fit_df, covid_norm_df)

## Split the data frame into training and test data sets
covid_split_df <- split_data(covid_model_df)

## Fit models based on the default list of machine learning (ML) algorithms
covid_model_list <- train_models(covid_split_df)

## Generate box plots to visualize performance of different ML algorithms
performance_plot(covid_model_list)

## Generate dot plots
performance_plot(covid_model_list, type = "dot")

## Change color palette
performance_plot(covid_model_list, type = "dot", palette = "inferno")
}