Normalize intensity data

This function normalizes data using a user-specified normalization method.

Usage

normalize_data(imp_df, method = "quantile")

Arguments

imp_df

An imp_df object with missing values imputed using impute_na.

method

Name of the normalization method to use. Choices are "none", "scale", "quantile" or "cyclicloess." Default is "quantile."

Value

A norm_df object, which is a data frame of normalized protein intensities.

Details

This function normalizes intensity values to achieve consistency among samples.

It assumes that the intensities in the data frame have been log-transformed, therefore, it is important to make sure that create_df was run with log_tr = TRUE(default) when creating the raw_df object.

See also

Author

Chathurani Ranathunge

Examples

if (FALSE) {
## Generate a raw_df object with default settings. No technical replicates.
raw_df <- create_df(
prot_groups = "https://raw.githubusercontent.com/caranathunge/promor_example_data/main/pg1.txt",
exp_design = "https://raw.githubusercontent.com/caranathunge/promor_example_data/main/ed1.txt"
)

## Impute missing values in the data frame using the default minProb
## method.
imp_df <- impute_na(raw_df)

## Normalize the imp_df object using the default quantile method
norm_df1 <- normalize_data(imp_df)

## Use the cyclicloess method
norm_df2 <- normalize_data(imp_df, method = "cyclicloess")
}